requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.5
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
